fix #66 : Icon config 분리
